requirements.txt
psycopg2-binary==2.9.7
Flask~=2.3.2
requests~=2.31.0
python-dotenv~=1.0.0

